refactor(WarekiPicker): client専用処理をclient/に切り出す (#7059)
