+++ packages/smarthr-ui/src/components/WarekiPicker/client/WarekiPicker.test.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react'
 import { act } from 'react'
 
-import { IntlProvider } from '../../intl'
-import { FormControl } from '../FormGroup'
+import { IntlProvider } from '../../../intl'
+import { FormControl } from '../../FormGroup'
 
 import { WarekiPicker } from './WarekiPicker'
