+++ packages/smarthr-ui/src/components/WarekiPicker/client/index.ts
@@ -1,0 +1,1 @@
+export { WarekiPicker } from './WarekiPicker'
