+++ packages/smarthr-ui/src/components/WarekiPicker/index.ts
@@ -1,1 +1,1 @@
-export { WarekiPicker } from './WarekiPicker'
+export { WarekiPicker } from './client'
