+++ packages/smarthr-ui/src/components/WarekiPicker/stories/VRTWarekiPicker.stories.tsx
@@ -2,7 +2,7 @@ import dayjs from 'dayjs'
 import { userEvent, within } from 'storybook/test'
 
 import { Cluster } from '../../Layout'
-import { WarekiPicker } from '../WarekiPicker'
+import { WarekiPicker } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
