+++ packages/smarthr-ui/src/components/WarekiPicker/stories/WarekiPicker.stories.tsx
@@ -1,7 +1,7 @@
 import dayjs from 'dayjs'
 import { userEvent, within } from 'storybook/test'
 
-import { WarekiPicker } from '../WarekiPicker'
+import { WarekiPicker } from '../client'
 
 import type { Meta, StoryObj } from '@storybook/react-vite'
